tools/kafka-services/mods_models/requirements.txt
numpy<1.19.0,>=1.16.0
pandas==1.1.4
tensorflow-gpu==2.3.1
keras==2.4.3
joblib==0.17.0
scikit-learn==0.23.2
mxnet==1.7.0.post1
mxnet-cu101==1.7.0
elasticsearch
es_pandas
kafka-python
pandas
progressbar2
